+++ CHANGELOG.md
@@ -21,6 +21,7 @@ $ npm install @ngxs/store@dev
 - Refactor(form-plugin): Replace `takeUntil` with `takeUntilDestroyed` [#2283](https://github.com/ngxs/store/pull/2283)
 - Refactor(router-plugin): Reduce RxJS depedency [#2291](https://github.com/ngxs/store/pull/2291)
 - Refactor(router-plugin): Change `@Selector` to `createSelector` [#2294](https://github.com/ngxs/store/pull/2294)
+- CI(bundlemon): Upgrade to bundlemon v3 [#2330](https://github.com/ngxs/store/pull/2330)
 
 ### 19.0.0 2024-12-3